Adding FixedKeysMapFactory
Not sure if this is just premature optimization, but this is an idea
for more-efficiently storing maps with a small number of fixed keys
(i.e. like the INFO or FORMAT fields of a VCF record). This uses
the same strategy as an EnumMap, but with a set of keys defined at
runtime.