Switched to XSSF
This makes the jar file a lot bigger (adding groovy went 2.2->6.1mb,
now adding groovy + poi-ooxml makes 15mb). But compared to the size
of the datasets... and the .xlsx files are smaller than .xls, so the
size gets amortized away. It may be nice to separate out the xls-making
part at some point, though.